requirements.txt
pyserial==3.5
mido==1.3.3